requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==4.0
django-environ==0.11.2
gunicorn==23.0.0
idna==3.10
oauthlib==3.2.2
packaging==24.1
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python-social-auth==0.3.6
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
tzdata==2024.2
urllib3==2.2.3
python-dotenv==1.0.1
psycopg2-binary==2.9.10
djangorestframework>=3.12,<4.0
django-keycloak-auth>=0.1.0
django-keycloak
